FEAT: Do some refactor
