+++ include/internal/internal.h
@@ -5,6 +5,10 @@
 #include <vector>
 #include <cstdint>
 
+#define KEY_FILE_VERSION 4U
+#define KEY_FILE_DERIVATION_FUNCTION std::string("scrypt")
+#define KEY_FILE_CIPHER_ALGORITHM std::string("aes-128-ctr")
+
 typedef std::vector<uint8_t> bytes;
 
 struct KdfParams
@@ -14,6 +18,20 @@ struct KdfParams
     uint32_t r = 8;
     uint32_t p = 1;
     std::string salt;
+};
+
+struct EncryptedData
+{
+    uint32_t version;
+    std::string id;
+    std::string address;
+    std::string bech32;
+    std::string kdf;
+    KdfParams kdfparams;
+    std::string iv;
+    std::string cipher;
+    std::string cipherText;
+    std::string mac;
 };
 
 #endif //ERD_INTERNAL_H
