+++ src/filehandler/keyfilereader.cpp
@@ -1,14 +1,12 @@
 #include "filehandler/keyfilereader.h"
 #include "cryptosignwrapper.h"
-
+#include "json/json.hpp"
 #include "errors.h"
-
 #include "hex.h"
 
 #include <fstream>
-#include <iostream>
 #include <stdexcept>
-#include "json/json.hpp"
+
 
 namespace internal
 {
@@ -29,6 +27,27 @@ bytes deriveSecretKey(EncryptedData const &data, std::string const& password)
 
    return wrapper::crypto::aes128ctrDecrypt(derivedKeyFirstHalf, data.cipherText, data.iv);
 }
+
+template <typename T>
+void checkParam(T const &param, T const &expectedParam, errorMessage const &error)
+{
+    if (param != expectedParam)
+    {
+        throw std::invalid_argument
+                (error + "Expected: " + std::to_string(expectedParam) +", got: " + std::to_string(param));
+    }
+}
+
+template <>
+void checkParam<std::string>(std::string const &param, std::string const &expectedParam, errorMessage const &error)
+{
+    if (param != expectedParam)
+    {
+        throw std::invalid_argument
+                (error + "Expected: " + expectedParam +", got: " + param);
+    }
+}
+
 }
 
 KeyFileReader::KeyFileReader(std::string const &filePath, std::string const &password) :
@@ -39,6 +58,11 @@ KeyFileReader::KeyFileReader(std::string const &filePath, std::string const &pas
         KeyFileReader::checkFile();
 
         auto const data = getFileContent();
+
+        internal::checkParam(data.version, KEY_FILE_VERSION, ERROR_MSG_KEY_FILE_VERSION);
+        internal::checkParam(data.cipher, KEY_FILE_CIPHER_ALGORITHM, ERROR_MSG_KEY_FILE_CIPHER);
+        internal::checkParam(data.kdf, KEY_FILE_DERIVATION_FUNCTION, ERROR_MSG_KEY_FILE_DERIVATION_FUNCTION);
+
         m_secretKey = internal::deriveSecretKey(data, password);
     }
     catch (std::exception const &error)
@@ -80,11 +104,16 @@ EncryptedData KeyFileReader::getFileContent() const
         std::ifstream stream(IFile::getFilePath());
         auto json = nlohmann::json::parse(stream);
 
+        data.version = json["version"];
+        data.kdf = json["crypto"]["kdf"];
+        data.cipher = json["crypto"]["cipher"];
+
         data.kdfparams.dklen = json["crypto"]["kdfparams"]["dklen"];
         data.kdfparams.n = json["crypto"]["kdfparams"]["n"];
         data.kdfparams.r = json["crypto"]["kdfparams"]["r"];
         data.kdfparams.p = json["crypto"]["kdfparams"]["p"];
         data.kdfparams.salt = util::hexToString(json["crypto"]["kdfparams"]["salt"]);
+
         data.iv = util::hexToString(json["crypto"]["cipherparams"]["iv"]);
         data.cipherText = util::hexToString(json["crypto"]["ciphertext"]);
         data.mac = util::hexToString(json["crypto"]["mac"]);
